+++ CHANGELOG.md
@@ -4,7 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Change-log](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## 7.0 (Stable, 2020-05-08)
+## 7.1.0 (Stable, 2020-05-15)
+- Added property for truncateLength on `vgr-card-section`
+
+## 7.0.0 (2020-05-08)
 - *Breaking change* `vgr-input` is updated, can now handle prefix and API more closely resembles HTML-input.
 - `vgr-list-column-checkbox` & `vgr-list-column-trashcan` are now officially removed.
 - *Deprecated* `vgr-filter-textbox` has been deprecated. `<vgr input type="search">` instead.